I CREATED UNIT TESTS FOR THE DATETIMEDISPLAY COMP
I decided to add Jest / Enzyme tests for the DateTimeDisplay component.

It is a good thing I did, because I found some bugs in the component.

I add the spec folder and the test file, and fixed the bugs in my
DateTimeDisplay component.

I also renamed the DateTimeDisplay.jsx file because I misspelled it as
DateTimDisplay.jsx.

@ralph-francesconi007
@robrichardsdpt
@nugrezo
@scottlurowist