I MADE SMALL NAV CHANGES IN PREP FOR RELEASE
Rob asked me this morning to change the navigation so that the SignIn
page is the landing page. He then asked me to make the NewsFeed
component the first page you see once you are logged in.

These required one line changes to App.js for the landing page, and
to the signin page success handler to navigate to the news feed page

@scottlurowist
@robrichardsdpt